added new advanced and beginner forms
you'll need to manually add a '2' to the old advanced and beginner.html
pages to access the new ones.  but the advanced is an ajax-enabled
search that filters drinks by name.  the beginner is a yet-to-be
complete 4-question form that returns results in the same fashion as
the advanced form.